requirements.txt
aiohttp==3.9.5
aiohttp==3.8.4
Pillow==9.4.0
Pillow==10.4.0
python-dotenv==1.0.1
qrcode==7.4.2
Requests==2.32.3
